refactor(orchestration): consolidate artifact handling into owner module
Move artifact storage, spill trigger, pointer render, and the read
tool from persistence.rs, orchestrator.rs, context/evidence.rs, and
tools/read_artifact.rs into one owner module at
persistence/artifacts/. persistence.rs stays as a facade; the read
tool file re-exports from the owner module. Behavior on every S2
manifest surface is byte-identical (852 lib tests pass under
INSTA_UPDATE=no).

Card: S5
