+++ crates/aura/src/orchestration/persistence.rs
@@ -1,81 +1,23 @@
 //! Execution persistence for orchestration observability.
 //!
-//! Writes detailed execution artifacts to disk asynchronously for debugging,
-//! analysis, and future retry intelligence. Supports iteration tracking for
-//! replanning scenarios.
-//!
-//! ## Directory Structure
-//!
-//! With session_id (web server path):
-//! ```text
-//! {base_path}/{session_id}/
-//! ├── latest -> {run_id}/              # Symlink to most recent run in session
-//! └── {run_id}/
-//!     ├── manifest.json                # Typed run manifest (RunManifest)
-//!     ├── artifacts/                   # Run-level result artifacts
-//!     │   └── task-0-default-iter-1-result.txt
-//!     └── iteration-{n}/              # One flat dir per iteration
-//!         ├── plan.json
-//!         ├── ...
-//! ```
-//!
-//! Without session_id (CLI/test path):
-//! ```text
-//! {base_path}/
-//! ├── latest -> {run_id}/
-//! └── {run_id}/
-//!     ├── manifest.json
-//!     ├── artifacts/
-//!     └── iteration-{n}/
-//! ```
+//! Run manifest schema, session history loading, and the artifact runtime
+//! facade. The artifact runtime lives in the `artifacts` submodule and is
+//! re-exported here so existing import paths remain unchanged.
 
 use serde::{Deserialize, Serialize};
 use std::io;
-use std::path::{Path, PathBuf};
-use std::sync::Arc;
-use std::sync::atomic::{AtomicUsize, Ordering};
-use std::time::Duration;
+use std::path::Path;
 use tokio::fs;
-use tokio::sync::{Mutex, MutexGuard, Notify};
-use tracing::Instrument;
 
 use super::events::RoutingMode;
-use super::types::{Plan, TaskStatus};
+use super::types::TaskStatus;
 
-// ============================================================================
-// Filename Helpers
-// ============================================================================
+pub mod artifacts;
 
-/// Sanitize a string for use as a filename component.
-///
-/// Lowercases, replaces non-alphanumeric characters with `-`, collapses
-/// consecutive `-`, and trims leading/trailing `-`. Returns `"unknown"` for
-/// empty input. Used for worker names and tool names in artifact filenames.
-pub fn sanitize_filename_component(s: &str) -> String {
-    let s = s.to_lowercase();
-    let sanitized: String = s
-        .chars()
-        .map(|c| if c.is_ascii_alphanumeric() { c } else { '-' })
-        .collect();
-    let collapsed = sanitized
-        .split('-')
-        .filter(|part| !part.is_empty())
-        .collect::<Vec<_>>()
-        .join("-");
-    if collapsed.is_empty() {
-        "unknown".to_string()
-    } else {
-        collapsed
-    }
-}
-
-/// True when `s` is safe to use as a single path component: non-empty, no path
-/// separators, no parent references. Artifact filenames and run IDs come from
-/// untrusted tool/LLM input and are validated with this before being joined
-/// into a persistence path.
-fn is_safe_path_component(s: &str) -> bool {
-    !s.is_empty() && !s.contains('/') && !s.contains('\\') && !s.contains("..")
-}
+pub use artifacts::{
+    ArtifactRef, ExecutionPersistence, SpilledArtifact, artifact_kind_from_filename,
+    lock_persistence, maybe_spill_result, sanitize_filename_component,
+};
 
 // ============================================================================
 // Run Manifest Types
@@ -272,733 +214,6 @@ pub struct TaskExecutionRecord {
     pub confidence: Option<String>,
     /// Notes for orchestrator (retry hints, blockers, etc.)
     pub orchestrator_notes: Option<String>,
-}
-
-/// Manages execution artifact persistence (async).
-///
-/// Tracks in-flight async writes via `in_flight` / `drain_notify` so callers
-/// can wait for all fire-and-forget `on_complete` persistence hooks to finish
-/// before reading back artifacts. The `Arc` counter and notify fields live
-/// outside the Mutex so increment/decrement is lock-free, but `on_complete`
-/// still acquires the Mutex for the actual file I/O — callers must release
-/// the Mutex before calling `drain()` to avoid deadlock.
-#[derive(Clone)]
-pub struct ExecutionPersistence {
-    base_path: PathBuf,
-    run_id: String,
-    session_id: Option<String>,
-    current_iteration: usize,
-    enabled: bool,
-    in_flight: Arc<AtomicUsize>,
-    drain_notify: Arc<Notify>,
-}
-
-impl ExecutionPersistence {
-    /// Create new persistence manager with unique run ID.
-    ///
-    /// Creates the run directory and a `latest` symlink.
-    ///
-    /// When `session_id` is provided, the directory structure becomes
-    /// `{base_path}/{session_id}/{run_id}/...`, grouping runs by session.
-    /// Without a session_id, the flat `{base_path}/{run_id}/...` layout is used.
-    pub async fn new<P: AsRef<Path>>(base_path: P, session_id: Option<String>) -> io::Result<Self> {
-        let base_path = base_path.as_ref().to_path_buf();
-
-        // Validate session_id to prevent path traversal
-        if let Some(ref sid) = session_id
-            && (sid.is_empty() || sid.contains('/') || sid.contains('\\') || sid.contains(".."))
-        {
-            return Err(io::Error::new(
-                io::ErrorKind::InvalidInput,
-                format!("Invalid session_id for persistence path: {:?}", sid),
-            ));
-        }
-
-        // Compute effective base: with session namespace or flat
-        let effective_base = if let Some(ref sid) = session_id {
-            base_path.join(sid)
-        } else {
-            base_path.clone()
-        };
-
-        // Generate unique run ID
-        let run_id = uuid::Uuid::now_v7().to_string();
-        let run_path = effective_base.join(&run_id);
-
-        fs::create_dir_all(&run_path).await?;
-
-        // Create symlink to latest run (best effort, ignore errors)
-        let latest_path = effective_base.join("latest");
-        let _ = tokio::fs::remove_file(&latest_path).await;
-
-        #[cfg(unix)]
-        {
-            use std::os::unix::fs::symlink;
-            let _ = tokio::task::spawn_blocking({
-                let run_id = run_id.clone();
-                let latest_path = latest_path.clone();
-                move || symlink(&run_id, latest_path)
-            })
-            .await;
-        }
-
-        tracing::info!(
-            "🗂️ Execution persistence initialized: {}",
-            run_path.display()
-        );
-
-        Ok(Self {
-            base_path: run_path,
-            run_id,
-            session_id,
-            current_iteration: 1,
-            enabled: true,
-            in_flight: Arc::new(AtomicUsize::new(0)),
-            drain_notify: Arc::new(Notify::new()),
-        })
-    }
-
-    /// Prune oldest run directories if the session exceeds `max_runs`.
-    ///
-    /// Skips the current run and the `latest` symlink. Directories are sorted
-    /// lexicographically (UUID v7 = chronological order) and the oldest are
-    /// removed first. Best-effort: errors on individual deletions are logged
-    /// but don't fail the operation.
-    pub async fn prune_session_runs(&self, max_runs: usize) {
-        if !self.enabled || max_runs == 0 || self.session_id.is_none() {
-            return;
-        }
-
-        let session_dir = match self.base_path.parent() {
-            Some(p) => p.to_path_buf(),
-            None => return,
-        };
-
-        let mut run_dirs: Vec<String> = Vec::new();
-        let mut entries = match fs::read_dir(&session_dir).await {
-            Ok(e) => e,
-            Err(_) => return,
-        };
-
-        while let Ok(Some(entry)) = entries.next_entry().await {
-            let path = entry.path();
-            if !path.is_dir() {
-                continue;
-            }
-            if let Some(name) = path.file_name().and_then(|n| n.to_str()) {
-                if name == "latest" || name == self.run_id {
-                    continue;
-                }
-                run_dirs.push(name.to_string());
-            }
-        }
-
-        if run_dirs.len() < max_runs {
-            return;
-        }
-
-        run_dirs.sort();
-        let to_remove = run_dirs.len() - max_runs + 1;
-        for dir_name in run_dirs.iter().take(to_remove) {
-            let path = session_dir.join(dir_name);
-            match fs::remove_dir_all(&path).await {
-                Ok(()) => tracing::info!("Pruned old run directory: {}", dir_name),
-                Err(e) => tracing::warn!("Failed to prune run directory {}: {}", dir_name, e),
-            }
-        }
-    }
-
-    /// Create a disabled persistence manager (no-op writes).
-    pub fn disabled() -> Self {
-        Self {
-            base_path: PathBuf::new(),
-            run_id: uuid::Uuid::new_v4().to_string(),
-            session_id: None,
-            current_iteration: 1,
-            enabled: false,
-            in_flight: Arc::new(AtomicUsize::new(0)),
-            drain_notify: Arc::new(Notify::new()),
-        }
-    }
-
-    /// Whether persistence is enabled (writes go to disk).
-    pub fn is_enabled(&self) -> bool {
-        self.enabled
-    }
-
-    /// Get the run ID for this execution.
-    pub fn run_id(&self) -> &str {
-        &self.run_id
-    }
-
-    /// Get the base path for this run's artifacts.
-    pub fn run_path(&self) -> &Path {
-        &self.base_path
-    }
-
-    /// Get current iteration number.
-    pub fn current_iteration(&self) -> usize {
-        self.current_iteration
-    }
-
-    /// Start a new iteration (for replanning).
-    pub fn start_new_iteration(&mut self) -> usize {
-        self.current_iteration += 1;
-        self.current_iteration
-    }
-
-    /// Arc handle to the in-flight write counter.
-    ///
-    /// Shared with `PersistenceWrapper` instances so fire-and-forget
-    /// `on_complete` hooks can increment/decrement without holding the Mutex.
-    pub fn in_flight_counter(&self) -> Arc<AtomicUsize> {
-        self.in_flight.clone()
-    }
-
-    /// Arc handle to the drain notification channel.
-    pub fn drain_notify(&self) -> Arc<Notify> {
-        self.drain_notify.clone()
-    }
-
-    /// Wait for all in-flight persistence writes to complete, bounded by `timeout`.
-    ///
-    /// Returns `true` if the counter reached zero before the deadline.
-    #[tracing::instrument(
-        name = "persistence.drain",
-        skip(self, timeout),
-        fields(
-            timeout_ms = timeout.as_millis() as u64,
-            remaining = tracing::field::Empty,
-        )
-    )]
-    pub async fn drain(&self, timeout: Duration) -> bool {
-        // Yield to let recently-spawned on_complete tasks poll their first
-        // increment before we check the counter (closes TOCTOU window between
-        // tokio::spawn and fetch_add inside on_complete).
-        tokio::task::yield_now().await;
-
-        if self.in_flight.load(Ordering::Acquire) == 0 {
-            return true;
-        }
-        tokio::select! {
-            _ = async {
-                while self.in_flight.load(Ordering::Acquire) > 0 {
-                    self.drain_notify.notified().await;
-                }
-            } => true,
-            _ = tokio::time::sleep(timeout) => {
-                let remaining = self.in_flight.load(Ordering::Acquire);
-                tracing::Span::current().record("remaining", remaining as i64);
-                tracing::warn!(remaining, "Persistence drain timed out");
-                false
-            }
-        }
-    }
-
-    /// Get iteration directory path (flat, directly under run dir).
-    pub(super) fn iteration_path(&self) -> PathBuf {
-        self.base_path
-            .join(format!("iteration-{}", self.current_iteration))
-    }
-
-    /// Build a dot-namespaced filename for a task attempt artifact.
-    fn task_attempt_filename(&self, task_id: usize, attempt: usize, suffix: &str) -> String {
-        format!("task-{}.attempt-{}.{}", task_id, attempt, suffix)
-    }
-
-    /// Write the plan created by coordinator.
-    #[tracing::instrument(
-        name = "persistence.write_plan",
-        skip(self, plan),
-        fields(
-            iteration = self.current_iteration,
-            task_count = plan.tasks.len(),
-        )
-    )]
-    pub async fn write_plan(&self, plan: &Plan) -> io::Result<PathBuf> {
-        if !self.enabled {
-            return Ok(PathBuf::new());
-        }
-
-        let iter_path = self.iteration_path();
-        fs::create_dir_all(&iter_path).await?;
-
-        let plan_path = iter_path.join("plan.json");
-        let json = serde_json::to_string_pretty(plan)
-            .map_err(|e| io::Error::new(io::ErrorKind::InvalidData, e))?;
-        fs::write(&plan_path, json).await?;
-
-        tracing::debug!("Written plan to: {}", plan_path.display());
-        Ok(plan_path)
-    }
-
-    /// Write planning phase artifacts (coordinator prompt/response).
-    #[tracing::instrument(
-        name = "persistence.write_planning_phase",
-        skip(self, prompt, response),
-        fields(
-            iteration = self.current_iteration,
-            prompt_bytes = prompt.len(),
-            response_bytes = response.len(),
-        )
-    )]
-    pub async fn write_planning_phase(&self, prompt: &str, response: &str) -> io::Result<PathBuf> {
-        if !self.enabled {
-            return Ok(PathBuf::new());
-        }
-
-        let iter_path = self.iteration_path();
-        fs::create_dir_all(&iter_path).await?;
-
-        fs::write(iter_path.join("planning.prompt.txt"), prompt).await?;
-        fs::write(iter_path.join("planning.response.txt"), response).await?;
-
-        Ok(iter_path)
-    }
-
-    /// Write worker task execution artifacts.
-    #[tracing::instrument(
-        name = "persistence.write_task_execution",
-        skip(self, prompt, response, record),
-        fields(
-            task_id,
-            attempt,
-            iteration = self.current_iteration,
-            prompt_bytes = prompt.len(),
-            response_bytes = response.len(),
-        )
-    )]
-    pub async fn write_task_execution(
-        &self,
-        task_id: usize,
-        attempt: usize,
-        prompt: &str,
-        response: &str,
-        record: &TaskExecutionRecord,
-    ) -> io::Result<PathBuf> {
-        if !self.enabled {
-            return Ok(PathBuf::new());
-        }
-
-        let iter_path = self.iteration_path();
-        fs::create_dir_all(&iter_path).await?;
-
-        // Write prompt and response with namespaced filenames.
-        // Tool calls are persisted incrementally via `append_tool_call()` to
-        // separate `*.tool-calls.json` files; nothing to write here.
-        let prompt_file = self.task_attempt_filename(task_id, attempt, "prompt.txt");
-        let response_file = self.task_attempt_filename(task_id, attempt, "response.txt");
-        fs::write(iter_path.join(&prompt_file), prompt).await?;
-        fs::write(iter_path.join(&response_file), response).await?;
-
-        // Write full execution record
-        let record_json = serde_json::to_string_pretty(record)
-            .map_err(|e| io::Error::new(io::ErrorKind::InvalidData, e))?;
-        let result_file = self.task_attempt_filename(task_id, attempt, "result.json");
-        fs::write(iter_path.join(&result_file), record_json).await?;
-
-        tracing::debug!(
-            "Written task execution to: {}/{}",
-            iter_path.display(),
-            prompt_file
-        );
-        Ok(iter_path)
-    }
-
-    /// Get relative path for logging.
-    pub fn relative_path(&self, task_id: usize, attempt: usize) -> String {
-        self.task_attempt_filename(task_id, attempt, "*")
-    }
-
-    // ========================================================================
-    // Result Artifact Methods
-    // ========================================================================
-
-    /// Directory for result artifacts (run-level, not per-iteration).
-    pub fn artifacts_path(&self) -> PathBuf {
-        self.base_path.join("artifacts")
-    }
-
-    /// Write a large result to an artifact file.
-    ///
-    /// Returns the artifact filename (not the full path) for reference in summaries.
-    /// Filenames are iteration-namespaced to avoid collisions across replans:
-    /// `task-{id}-{worker}-iter-{n}-result.txt`
-    #[tracing::instrument(
-        name = "persistence.write_result_artifact",
-        skip(self, result, worker_name),
-        fields(
-            task_id,
-            iteration,
-            worker = worker_name.unwrap_or("default"),
-            result_bytes = result.len(),
-        )
-    )]
-    pub async fn write_result_artifact(
-        &self,
-        task_id: usize,
-        worker_name: Option<&str>,
-        iteration: usize,
-        result: &str,
-    ) -> io::Result<String> {
-        if !self.enabled {
-            return Ok(String::new());
-        }
-
-        let artifacts_dir = self.artifacts_path();
-        fs::create_dir_all(&artifacts_dir).await?;
-
-        let worker = sanitize_filename_component(worker_name.unwrap_or("default"));
-        let filename = format!("task-{}-{}-iter-{}-result.txt", task_id, worker, iteration);
-        let artifact_path = artifacts_dir.join(&filename);
-        fs::write(&artifact_path, result).await?;
-
-        tracing::info!(
-            "Written result artifact ({} chars) to: {}",
-            result.len(),
-            artifact_path.display()
-        );
-        Ok(filename)
-    }
-
-    /// Write a tool output to an artifact file.
-    ///
-    /// Returns the artifact filename for reference in footers and ToolCallRecord.
-    /// Filename: `task-{id}-{worker}-iter-{n}-{tool_name}-{call_idx}-output.txt`
-    #[tracing::instrument(
-        name = "persistence.write_tool_output_artifact",
-        skip(self, output),
-        fields(
-            task_id,
-            worker_name,
-            iteration,
-            tool_name,
-            call_idx,
-            output_bytes = output.len(),
-        )
-    )]
-    pub async fn write_tool_output_artifact(
-        &self,
-        task_id: usize,
-        worker_name: &str,
-        iteration: usize,
-        tool_name: &str,
-        call_idx: usize,
-        output: &str,
-    ) -> io::Result<String> {
-        if !self.enabled {
-            return Ok(String::new());
-        }
-
-        let artifacts_dir = self.artifacts_path();
-        fs::create_dir_all(&artifacts_dir).await?;
-
-        let worker = sanitize_filename_component(worker_name);
-        let tool = sanitize_filename_component(tool_name);
-        let filename = format!(
-            "task-{}-{}-iter-{}-{}-{}-output.txt",
-            task_id, worker, iteration, tool, call_idx
-        );
-        let artifact_path = artifacts_dir.join(&filename);
-        fs::write(&artifact_path, output).await?;
-
-        tracing::info!(
-            "Written tool output artifact ({} chars) to: {}",
-            output.len(),
-            artifact_path.display()
-        );
-        Ok(filename)
-    }
-
-    /// Read an artifact file by filename.
-    ///
-    /// Resolves the path via [`artifact_path`](Self::artifact_path) (which
-    /// validates against path traversal), then reads it.
-    pub async fn read_artifact(&self, filename: &str) -> io::Result<String> {
-        if !self.enabled {
-            return Err(io::Error::new(
-                io::ErrorKind::NotFound,
-                "Persistence is disabled",
-            ));
-        }
-        let artifact_path = self.artifact_path(filename)?;
-        fs::read_to_string(&artifact_path).await
-    }
-
-    /// Read an artifact from a different run in the same session.
-    ///
-    /// Resolves the path via
-    /// [`artifact_path_cross_run`](Self::artifact_path_cross_run), then adds a
-    /// canonicalized containment check (the resolved path must stay under the
-    /// session directory) before reading.
-    pub async fn read_artifact_cross_run(
-        &self,
-        filename: &str,
-        run_id: &str,
-    ) -> io::Result<String> {
-        if !self.enabled {
-            return Err(io::Error::new(
-                io::ErrorKind::NotFound,
-                "Persistence is disabled",
-            ));
-        }
-
-        let artifact_path = self.artifact_path_cross_run(filename, run_id)?;
-
-        // Defense-in-depth: the resolved path must canonicalize to somewhere
-        // under the session directory (catches symlink escapes the lexical
-        // component checks in artifact_path_cross_run can't see).
-        let session_dir = self
-            .base_path
-            .parent()
-            .ok_or_else(|| io::Error::new(io::ErrorKind::NotFound, "No parent directory"))?;
-        let canonical_session = session_dir
-            .canonicalize()
-            .unwrap_or_else(|_| session_dir.to_path_buf());
-        let canonical_artifact = artifact_path
-            .canonicalize()
-            .map_err(|e| io::Error::new(e.kind(), format!("Artifact not found: {e}")))?;
-        if !canonical_artifact.starts_with(&canonical_session) {
-            return Err(io::Error::new(
-                io::ErrorKind::InvalidInput,
-                "Cross-run artifact path escapes session directory",
-            ));
-        }
-
-        fs::read_to_string(&artifact_path).await
-    }
-
-    /// Resolve the absolute path of a current-run artifact.
-    ///
-    /// Validates `filename` against path traversal but does **not** check that
-    /// the file exists or read it. Used both by [`read_artifact`](Self::read_artifact)
-    /// and to build an in-place `file=` reference for the scratchpad read tools
-    /// when an artifact is too large to inline.
-    pub fn artifact_path(&self, filename: &str) -> io::Result<PathBuf> {
-        if !is_safe_path_component(filename) {
-            return Err(io::Error::new(
-                io::ErrorKind::InvalidInput,
-                "Invalid artifact filename",
-            ));
-        }
-        Ok(self.artifacts_path().join(filename))
-    }
-
-    /// Resolve the absolute path of a cross-run artifact within this session.
-    ///
-    /// Resolves against `{session_dir}/{run_id}/artifacts/{filename}` where
-    /// `session_dir` is the parent of the current run directory. Validates both
-    /// components against path traversal but does **not** check existence; the
-    /// caller (e.g. [`read_artifact_cross_run`](Self::read_artifact_cross_run))
-    /// performs the canonicalized containment check when it reads.
-    pub fn artifact_path_cross_run(&self, filename: &str, run_id: &str) -> io::Result<PathBuf> {
-        if !is_safe_path_component(run_id) {
-            return Err(io::Error::new(
-                io::ErrorKind::InvalidInput,
-                "Invalid run_id for cross-run artifact read",
-            ));
-        }
-        if !is_safe_path_component(filename) {
-            return Err(io::Error::new(
-                io::ErrorKind::InvalidInput,
-                "Invalid artifact filename",
-            ));
-        }
-        let session_dir = self
-            .base_path
-            .parent()
-            .ok_or_else(|| io::Error::new(io::ErrorKind::NotFound, "No parent directory"))?;
-        Ok(session_dir.join(run_id).join("artifacts").join(filename))
-    }
-
-    /// List all artifact filenames.
-    pub async fn list_artifacts(&self) -> io::Result<Vec<String>> {
-        if !self.enabled {
-            return Ok(Vec::new());
-        }
-
-        let artifacts_dir = self.artifacts_path();
-        if !artifacts_dir.exists() {
-            return Ok(Vec::new());
-        }
-
-        let mut entries = fs::read_dir(&artifacts_dir).await?;
-        let mut filenames = Vec::new();
-        while let Some(entry) = entries.next_entry().await? {
-            if let Some(name) = entry.file_name().to_str() {
-                filenames.push(name.to_string());
-            }
-        }
-        filenames.sort();
-        Ok(filenames)
-    }
-
-    /// List all artifact filenames with file sizes.
-    pub async fn list_artifacts_with_metadata(&self) -> io::Result<Vec<(String, u64)>> {
-        if !self.enabled {
-            return Ok(Vec::new());
-        }
-
-        let artifacts_dir = self.artifacts_path();
-        if !artifacts_dir.exists() {
-            return Ok(Vec::new());
-        }
-
-        let mut entries = fs::read_dir(&artifacts_dir).await?;
-        let mut results = Vec::new();
-        while let Some(entry) = entries.next_entry().await? {
-            if let Some(name) = entry.file_name().to_str() {
-                let size = entry.metadata().await.map(|m| m.len()).unwrap_or(0);
-                results.push((name.to_string(), size));
-            }
-        }
-        results.sort_by(|a, b| a.0.cmp(&b.0));
-        Ok(results)
-    }
-
-    /// Load all tool call records for a given task across all iterations.
-    ///
-    /// Scans `iteration-*/task-{task_id}.attempt-*.tool-calls.json` under the
-    /// run directory. Returns an empty vec on file-not-found (graceful for old runs).
-    pub async fn load_tool_records_for_task(&self, task_id: usize) -> Vec<ToolCallRecord> {
-        if !self.enabled {
-            return Vec::new();
-        }
-
-        let mut all_records = Vec::new();
-        let prefix = format!("task-{task_id}.attempt-");
-
-        for iter_num in 1..=self.current_iteration {
-            let iter_dir = self.base_path.join(format!("iteration-{iter_num}"));
-            if !iter_dir.exists() {
-                continue;
-            }
-            let Ok(mut entries) = fs::read_dir(&iter_dir).await else {
-                continue;
-            };
-            while let Ok(Some(entry)) = entries.next_entry().await {
-                let name = entry.file_name();
-                let Some(name_str) = name.to_str() else {
-                    continue;
-                };
-                if !name_str.starts_with(&prefix) || !name_str.ends_with(".tool-calls.json") {
-                    continue;
-                }
-                let Ok(content) = fs::read_to_string(entry.path()).await else {
-                    continue;
-                };
-                if let Ok(records) = serde_json::from_str::<Vec<ToolCallRecord>>(&content) {
-                    all_records.extend(records);
-                }
-            }
-        }
-
-        all_records
-    }
-
-    // ========================================================================
-    // Run Manifest
-    // ========================================================================
-
-    /// Get the session ID (if set).
-    pub fn session_id(&self) -> Option<&str> {
-        self.session_id.as_deref()
-    }
-
-    /// Write a typed run manifest to `{run_path}/manifest.json`.
-    ///
-    /// Called at the end of `run_orchestration_loop()` on both success and
-    /// failure paths. The manifest serves as a structured index for Phase 2
-    /// cross-turn context.
-    #[tracing::instrument(
-        name = "persistence.write_manifest",
-        skip(self, manifest),
-        fields(
-            task_count = manifest.task_summaries.len(),
-            artifact_count = manifest.artifact_paths.len(),
-        )
-    )]
-    pub async fn write_manifest(&self, manifest: &RunManifest) -> io::Result<PathBuf> {
-        if !self.enabled {
-            return Ok(PathBuf::new());
-        }
-
-        let manifest_path = self.base_path.join("manifest.json");
-        let json = serde_json::to_string_pretty(manifest)
-            .map_err(|e| io::Error::new(io::ErrorKind::InvalidData, e))?;
-        fs::write(&manifest_path, json).await?;
-
-        tracing::info!("Written run manifest to: {}", manifest_path.display());
-        Ok(manifest_path)
-    }
-
-    /// Append a tool call record to the current task's execution.
-    ///
-    /// This is called by PersistenceWrapper during tool execution.
-    /// Tool calls are appended to a running list, not overwritten.
-    #[tracing::instrument(
-        name = "persistence.append_tool_call",
-        skip(self, record),
-        fields(
-            task_id,
-            attempt,
-            iteration = self.current_iteration,
-            tool = record.tool,
-            output_bytes = record.output.as_ref().map(|o| o.len()).unwrap_or(0),
-        )
-    )]
-    pub async fn append_tool_call(
-        &self,
-        task_id: usize,
-        attempt: usize,
-        record: &ToolCallRecord,
-    ) -> io::Result<()> {
-        if !self.enabled {
-            return Ok(());
-        }
-
-        let iter_path = self.iteration_path();
-        fs::create_dir_all(&iter_path).await?;
-
-        let tool_file = self.task_attempt_filename(task_id, attempt, "tool-calls.json");
-        let tool_calls_path = iter_path.join(&tool_file);
-
-        // Read existing tool calls or start fresh
-        let mut tool_calls: Vec<ToolCallRecord> = if tool_calls_path.exists() {
-            let content = fs::read_to_string(&tool_calls_path).await?;
-            serde_json::from_str(&content).unwrap_or_default()
-        } else {
-            Vec::new()
-        };
-
-        // Append new record
-        tool_calls.push(record.clone());
-
-        // Write back
-        let json = serde_json::to_string_pretty(&tool_calls)
-            .map_err(|e| io::Error::new(io::ErrorKind::InvalidData, e))?;
-        fs::write(&tool_calls_path, json).await?;
-
-        tracing::debug!(
-            "Appended tool call to: {} (total: {})",
-            tool_calls_path.display(),
-            tool_calls.len()
-        );
-
-        Ok(())
-    }
-}
-
-/// Acquire the `ExecutionPersistence` lock with a tracing span around the wait.
-///
-/// Every consumer should use this for lock-acquisition visibility; the span is
-/// named `persistence_lock` and carries the logical operation being performed.
-pub async fn lock_persistence<'a>(
-    mutex: &'a Arc<Mutex<ExecutionPersistence>>,
-    operation: &'static str,
-) -> MutexGuard<'a, ExecutionPersistence> {
-    let span = tracing::info_span!("persistence_lock", operation);
-    async { mutex.lock().await }.instrument(span).await
 }
 
 // ============================================================================
@@ -1210,6 +425,8 @@ fn render_task_summary(task: &TaskSummary, out: &mut String) {
 #[cfg(test)]
 mod tests {
     use super::*;
+    use std::sync::atomic::Ordering;
+    use std::time::Duration;
     use tempfile::TempDir;
 
     #[tokio::test]
@@ -1287,7 +504,9 @@ mod tests {
         assert!(!persistence.enabled);
 
         // All writes should succeed but do nothing
-        let result = persistence.write_plan(&Plan::new("test")).await;
+        let result = persistence
+            .write_plan(&super::super::types::Plan::new("test"))
+            .await;
         assert!(result.is_ok());
     }
 
@@ -1487,133 +706,6 @@ mod tests {
 
     // ========================================================================
     // Run Manifest Tests
-    // ========================================================================
-
-    #[tokio::test]
-    async fn test_manifest_serde_roundtrip() {
-        let manifest = RunManifest {
-            run_id: "test-run-id".to_string(),
-            session_id: Some("cs_abc".to_string()),
-            timestamp: "2026-03-19T12:00:00Z".to_string(),
-            goal: "Test the system".to_string(),
-            status: RunStatus::Success,
-            iterations: 2,
-            routing_mode: Some(RoutingMode::Orchestrated),
-            outcome: None,
-            response_summary: None,
-            task_summaries: vec![
-                TaskSummary {
-                    task_id: 0,
-                    description: "First task".to_string(),
-                    status: TaskStatus::Complete,
-                    worker: Some("research".to_string()),
-                    result_preview: Some("The answer is 42".to_string()),
-                    confidence: None,
-                    failure_category: None,
-                    error: None,
-                    error_context: None,
-                    tool_trace: vec![],
-                    artifacts: vec![],
-                },
-                TaskSummary {
-                    task_id: 1,
-                    description: "Second task".to_string(),
-                    status: TaskStatus::Failed,
-                    worker: None,
-                    result_preview: None,
-                    confidence: None,
-                    failure_category: Some(super::super::types::FailureCategory::AgentError),
-                    error: Some("Connection refused".to_string()),
-                    error_context: None,
-                    tool_trace: vec![],
-                    artifacts: vec![],
-                },
-            ],
-            artifact_paths: vec!["task-0-research-iter-1-result.txt".to_string()],
-        };
-
-        let json = serde_json::to_string_pretty(&manifest).unwrap();
-        let deserialized: RunManifest = serde_json::from_str(&json).unwrap();
-
-        assert_eq!(deserialized.run_id, "test-run-id");
-        assert_eq!(deserialized.session_id, Some("cs_abc".to_string()));
-        assert_eq!(deserialized.status, RunStatus::Success);
-        assert_eq!(deserialized.iterations, 2);
-        assert_eq!(deserialized.task_summaries.len(), 2);
-        assert_eq!(deserialized.task_summaries[0].status, TaskStatus::Complete);
-        assert_eq!(deserialized.task_summaries[1].status, TaskStatus::Failed);
-        assert_eq!(
-            deserialized.artifact_paths,
-            vec!["task-0-research-iter-1-result.txt"]
-        );
-    }
-
-    #[tokio::test]
-    async fn test_write_manifest() {
-        let temp_dir = TempDir::new().unwrap();
-        let persistence = ExecutionPersistence::new(temp_dir.path().join("memory"), None)
-            .await
-            .unwrap();
-
-        let manifest = RunManifest {
-            run_id: persistence.run_id().to_string(),
-            session_id: None,
-            timestamp: "2026-03-19T12:00:00Z".to_string(),
-            goal: "Test goal".to_string(),
-            status: RunStatus::PartialSuccess,
-            iterations: 1,
-            routing_mode: Some(RoutingMode::Routed),
-            outcome: None,
-            response_summary: None,
-            task_summaries: vec![],
-            artifact_paths: vec![],
-        };
-
-        let path = persistence.write_manifest(&manifest).await.unwrap();
-        assert!(path.exists());
-
-        // Read back and verify
-        let content = tokio::fs::read_to_string(&path).await.unwrap();
-        let read_back: RunManifest = serde_json::from_str(&content).unwrap();
-        assert_eq!(read_back.goal, "Test goal");
-        assert_eq!(read_back.status, RunStatus::PartialSuccess);
-    }
-
-    #[tokio::test]
-    async fn test_write_manifest_disabled() {
-        let persistence = ExecutionPersistence::disabled();
-        let manifest = RunManifest {
-            run_id: String::new(),
-            session_id: None,
-            timestamp: String::new(),
-            goal: String::new(),
-            status: RunStatus::Failed,
-            iterations: 0,
-            routing_mode: None,
-            outcome: None,
-            response_summary: None,
-            task_summaries: vec![],
-            artifact_paths: vec![],
-        };
-        let path = persistence.write_manifest(&manifest).await.unwrap();
-        assert_eq!(path, PathBuf::new());
-    }
-
-    #[tokio::test]
-    async fn test_run_status_serde() {
-        // Verify snake_case serialization
-        let json = serde_json::to_string(&RunStatus::PartialSuccess).unwrap();
-        assert_eq!(json, "\"partial_success\"");
-
-        let json = serde_json::to_string(&RunStatus::Success).unwrap();
-        assert_eq!(json, "\"success\"");
-
-        let json = serde_json::to_string(&RunStatus::Failed).unwrap();
-        assert_eq!(json, "\"failed\"");
-    }
-
-    // ========================================================================
-    // Session History Tests
     // ========================================================================
 
     fn make_test_manifest(run_id: &str, timestamp: &str, goal: &str) -> RunManifest {
